Move test files too
