+++ src/contentScript/pageEditor/selectElement.ts
@@ -1,3 +1,20 @@
+/*
+ * Copyright (C) 2024 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 import {
   findContainer,
   inferMultiElementSelector,
@@ -10,7 +27,7 @@ import inferSingleElementSelector from "@/utils/inference/inferSingleElementSele
 import { type ElementInfo } from "@/utils/inference/selectorTypes";
 import { userSelectElement } from "./elementPicker";
 
-export async function selectElement({
+export default async function selectElement({
   mode = "element",
   root,
   isMulti: initialIsMulti = false,
